requirements.txt
opencv-python
numpy
easyocr
pillow
pytesseract
PyQt5
